requirements.txt
fastapi
uvicorn[standard]
PyJWT
pydantic
requests
python-multipart
python-dateutil
billboard.py